git-intelligence/requirements.txt
GitPython>=3.1.0
click>=8.0.0
rich>=13.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
